move analytics into remote_logging
 * move download logging from analytics to remote_logging, and update it
   to log all action types.
