+++ waterbutler/server/api/v0/core.py
@@ -9,7 +9,6 @@ from raven.contrib.tornado import SentryMixin
 from waterbutler import tasks
 from waterbutler.core import utils
 from waterbutler.core import signing
-from waterbutler.core import analytics
 from waterbutler.core import exceptions
 from waterbutler.server import settings
 from waterbutler.core import remote_logging
@@ -98,15 +97,6 @@ class BaseProviderHandler(BaseHandler):
         remote_logging.log_file_action(action, source=source, api_version='v0',
                                        request=utils._serialize_request(self.request),
                                        size=self.bytes_written,)
-
-    @utils.async_retry(retries=5, backoff=5)
-    async def _log_download(self, action, metadata=None, path=None):
-        if action not in ('download_file', 'download_zip'):
-            return
-        downloadee = LogPayload(self.arguments['nid'], self.provider, metadata=metadata, path=path)
-        await analytics.log_download(action, payload=downloadee, api_version='v0',
-                                     request=utils._serialize_request(self.request),
-                                     size=self.bytes_written)
 
 
 class BaseCrossProviderHandler(BaseHandler):
