+++ waterbutler/server/api/v1/provider/__init__.py
@@ -6,7 +6,6 @@ import logging
 import tornado.gen
 
 from waterbutler.core import utils
-from waterbutler.core import analytics
 from waterbutler.server import settings
 from waterbutler.server.api.v1 import core
 from waterbutler.core import remote_logging
@@ -137,7 +136,6 @@ class ProviderHandler(core.BaseHandler, CreateMixin, MetadataMixin, MoveCopyMixi
         }[method]()
 
         self._send_hook(action)
-        self._log_downloads(action)
 
     def _send_hook(self, action):
         source = None
@@ -156,7 +154,7 @@ class ProviderHandler(core.BaseHandler, CreateMixin, MetadataMixin, MoveCopyMixi
             )
         elif action in ('create', 'create_folder', 'update'):
             source = LogPayload(self.resource, self.provider, metadata=self.metadata)
-        elif action in ('delete',):
+        elif action in ('delete', 'download_file', 'download_zip'):
             source = LogPayload(self.resource, self.provider, path=self.path)
         else:
             return
@@ -164,12 +162,3 @@ class ProviderHandler(core.BaseHandler, CreateMixin, MetadataMixin, MoveCopyMixi
         remote_logging.log_file_action(action, source=source, destination=destination, api_version='v1',
                                        request=utils._serialize_request(self.request),
                                        size=self.bytes_written,)
-
-    @utils.async_retry(retries=5, backoff=5)
-    async def _log_downloads(self, action):
-        if action not in ('download_file', 'download_zip'):
-            return
-        downloadee = LogPayload(self.resource, self.provider, path=self.path)
-        await analytics.log_download(action, payload=downloadee, api_version='v1',
-                                     request=utils._serialize_request(self.request),
-                                     size=self.bytes_written)
